Adjust ts.js to use tiddlyweb.status from /status.js
Instead making an async call to to /status. This is a first step
in lessening complexity in application code. See the statusjs
branch (forthcoming) in tiddlyspace for comparison.

Also, only do all that work for ts.initForSpace_ if we are in _space
or _account. For things like backstage it is not needed. Obviously
there are further ways to clean this up up, but I thought I'd
start with this for review.