When there is an error getting includes, report it correctly.
The previous code reported that only members could see includes.
This is not true. The public recipe is _public_, therefore anyone
can see it. Thus, if there is an error here it is going to be
something weird and it is best if we know the HTTP response code
so we can make some real diagnosis, or at least find the entry
in the log.