Extract message display utility functions.
ts.messages.display() does not provide additional value over
just displayMessage(). This is because ts.js is supposed to
provide standardized behavior to forms (as widgets) that are
used throughout tiddlyspace admin interfaces used by users.

That is: the message display functionality isn't going to be
exported; it is available as a result of form being ts.js-ified.

Again: small steps in achieving a form of scrutability in the code.