Move init handling that is done if status is present to own function.
In the process change loginStatus to use its own variables,
as they are not used anywhere else.

Remove the tsInitializing classes. These were only used because
status was being loaded asynch. Now, with status.js, by the time
this code is reached, we have status, so the changes are not relevant.