extract and rename loadHash
The name doesn't describe what it actually does. Now it does.

And it would be better as a straight function, so made it return
its results, rather than assign then within.

However, in order to make it testable, it was necessary to stick
the function back on the ts object so it can be reached.

The test is rudimentary, just as a proof. The complexity suggests
that the way in which the object is reached/established is a bit
wack.