Allow login to POST to custom challenger obtained from form
The ts.js has an options literal that can have a custom challenger set against it but this option is never utilised.

This commit allows the login code to obtain a custom challenger to use for login from the form by reading a hidden field value.  If it is not provided it will use the default tiddlyspace cookie form challenger.

This commit also reports on test coverage achieved so far.